+++ source/numberDetection/clipping_number.py
